split llm.py file into multiple files
