+++ backend/main.py
@@ -6,7 +6,7 @@ from fastapi import FastAPI
 
 from modal import Image, App, asgi_app, Secret
 from routers import workflow, chat, firmware, plugins, transcribe, notifications, \
-    speech_profile, agents, users, processing_conversations, trends, sync, apps, custom_auth, \
+    speech_profile, agents, users, trends, sync, apps, custom_auth, \
     payment, integration, conversations, memories, mcp
 
 from utils.other.timeout import TimeoutMiddleware
@@ -31,7 +31,6 @@ app.include_router(workflow.router)
 app.include_router(integration.router)
 app.include_router(agents.router)
 app.include_router(users.router)
-app.include_router(processing_conversations.router)
 app.include_router(trends.router)
 
 app.include_router(firmware.router)
