+++ backend/utils/social.py
@@ -9,7 +9,7 @@ from ulid import ULID
 from database.apps import update_app_in_db, upsert_app_to_db, get_persona_by_id_db, \
     get_persona_by_username_twitter_handle_db
 from database.redis_db import delete_generic_cache, save_username, is_username_taken
-from utils.llm import condense_tweets, generate_twitter_persona_prompt
+from utils.llm.persona import condense_tweets, generate_twitter_persona_prompt
 from utils.conversations.memories import process_twitter_memories
 
 rapid_api_host = os.getenv('RAPID_API_HOST')
